tree structure for artifacts crafting
